UNR-372 cleanup and rename (#18)
@Vatyx  Sorry for the GIGANTIC PR.

The first commit fixes a few bugs.

The rest of the commits do the following:
* Re-structure the Content/ and Source/ directories to make them more sensible
* Rename the project to ThirdPersonShooter
* Change all class prefixes to `TPS`
* Clean up all blueprints to reference the correct parent classes
* Rename blueprint classes too
* Fix remaining bugs

The PR in its current state should work, locally, but I haven't tested on the cloud or really stress tested it at all. It would be worth running through the standard game flows with James ( @InvalidEntry ) before releasing.